mysite/requirements.txt
# Dev environment
pip
autopep8

# App
asgiref==3.7.2
Django==5.0.13
sqlparse==0.4.4
typing_extensions==4.9.0
djangorestframework==3.16.0
firebase-admin==6.9.0
psycopg2-binary==2.9.10
django-filter==25.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
python-dotenv==1.1.1
gunicorn==23.0.0